fix: id change int to long

+++ app/src/main/java/com/exfarnanda1945/loremipsumgen/feat_history/domain/repository/IHistoryRepository.kt
@@ -4,7 +4,7 @@ import com.exfarnanda1945.loremipsumgen.core.utils.Resource
 import com.exfarnanda1945.loremipsumgen.feat_history.domain.models.HistoryGenerator
 
 interface IHistoryRepository {
-    suspend fun insert(data:HistoryGenerator):Resource<Int>
+    suspend fun insert(data:HistoryGenerator):Resource<Long>
     suspend fun lists():Resource<List<HistoryGenerator>>
     suspend fun get(id:Int):Resource<HistoryGenerator>
     suspend fun delete(id:Int):Resource<Unit>
